clean print-outs
