+++ submissions/simple_LSTM/classifier.py
@@ -2,7 +2,6 @@ from sklearn.ensemble import RandomForestClassifier
 from sklearn.decomposition import PCA
 from sklearn.pipeline import Pipeline
 from sklearn.base import BaseEstimator
-import numpy as np
 
 
 class Classifier(BaseEstimator):
