+++ submissions/simple_LSTM/feature_extractor_clf.py
