+++ submissions/tf_circle_fit/classifier.py
@@ -2,7 +2,6 @@ from sklearn.ensemble import RandomForestClassifier
 from sklearn.decomposition import PCA
 from sklearn.pipeline import Pipeline
 from sklearn.base import BaseEstimator
-import numpy as np
 
 
 class Classifier(BaseEstimator):
@@ -20,8 +19,10 @@ class Classifier(BaseEstimator):
 
     def predict(self, X):
         y_pred = self.clf.predict(X)
+        print("y_pred clf : ", y_pred)
         return y_pred
 
     def predict_proba(self, X):
         y_pred_proba = self.clf.predict_proba(X)
+        print("y_pred_proba clf : ", y_pred_proba)
         return y_pred_proba
