+++ submissions/tf_circle_fit/feature_extractor_clf.py
@@ -9,5 +9,7 @@ class FeatureExtractor():
         pass
 
     def transform(self, X_df):
+        print("X_df fe clf: ", X_df)
         X_array = X_df.values
+        print("X_array fe clf: ", X_array)
         return X_array
